# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend normativo de acceso a derechos",
    "description": "\nAPI de consulta del corpus normativo de acceso a derechos.\n\nToda respuesta lleva la misma envoltura: `schema_version`, `release_id`,\n`as_of`, `known_at`, `data_status`, `evidence`, `missing_fields` y `warnings`.\nSin `as_of` una respuesta no dice para cuándo vale; sin `release_id` no se puede\nreproducir; sin `missing_fields` una abstención sería indistinguible de un \"no\".\n\nLos errores son tipados. Información insuficiente es un estado de dominio y\nnunca se disfraza de 500.\n\nEsta API no otorga, no rechaza y no revoca prestaciones. Devuelve lo que dicen\nlas normas publicadas, con la evidencia que lo sostiene y lo que falta averiguar.\n",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/normas": {
      "get": {
        "tags": [
          "normas"
        ],
        "summary": "Listar Normas",
        "description": "Normas del corpus con su cobertura de campos.\n\nDevuelve normas identificadas aunque todavía no tengan versiones publicadas:\nsaber que una norma existe y que su ficha está incompleta es información\nútil, y ocultarla haría creer que no existe.",
        "operationId": "listar_normas_v1_normas_get",
        "parameters": [
          {
            "name": "texto",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Búsqueda textual en el título.",
              "title": "Texto"
            },
            "description": "Búsqueda textual en el título."
          },
          {
            "name": "jurisdiccion",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiccion"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "numero",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Numero"
            }
          },
          {
            "name": "anio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anio"
            }
          },
          {
            "name": "materia",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Materia"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_Pagina_NormaResumen__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/normas/{norma_id}": {
      "get": {
        "tags": [
          "normas"
        ],
        "summary": "Obtener Norma",
        "description": "Ficha de una norma: versiones, siete campos, relaciones y pendientes.\n\nCada versión declara si puede servirse para la capacidad consultada y, si no,\npor qué. Una abstención explicada es información; una lista vacía no.",
        "operationId": "obtener_norma_v1_normas__norma_id__get",
        "parameters": [
          {
            "name": "norma_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Norma Id"
            }
          },
          {
            "name": "capacidad",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Capacidad para la que se evalúa si cada versión puede servirse.",
              "default": "IDENTIFICACION",
              "title": "Capacidad"
            },
            "description": "Capacidad para la que se evalúa si cada versión puede servirse."
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_NormaDetalle_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/revisiones/{incidencia_id}/resolver": {
      "post": {
        "tags": [
          "administración"
        ],
        "summary": "Resolver Revision",
        "operationId": "resolver_revision_v1_admin_revisiones__incidencia_id__resolver_post",
        "parameters": [
          {
            "name": "incidencia_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incidencia Id"
            }
          },
          {
            "name": "X-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudResolucion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Resolver Revision V1 Admin Revisiones  Incidencia Id  Resolver Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/releases": {
      "post": {
        "tags": [
          "administración"
        ],
        "summary": "Crear Release",
        "operationId": "crear_release_v1_admin_releases_post",
        "parameters": [
          {
            "name": "X-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudRelease"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Crear Release V1 Admin Releases Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Advertencia": {
        "properties": {
          "codigo": {
            "$ref": "#/components/schemas/CodigoError"
          },
          "detalle": {
            "type": "string",
            "title": "Detalle"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "detalle"
        ],
        "title": "Advertencia"
      },
...
      "CodigoError": {
        "type": "string",
        "enum": [
          "SOURCE_UNAVAILABLE",
          "STALE_DATA",
          "CONFLICT",
          "INSUFFICIENT_EVIDENCE",
          "UNKNOWN_IDENTITY",
          "UNSUPPORTED_SCOPE",
          "INVALID_REQUEST",
          "NOT_AUTHORIZED",
          "VERSION_CONFLICT"
        ],
        "title": "CodigoError",
        "description": "Errores tipados del contrato.\n\nInformación insuficiente es un estado de dominio, no un fallo del servidor:\nnunca se disfraza de 500."
      },
...
      "DataStatus": {
        "type": "string",
        "enum": [
          "PUBLICADO",
          "PARCIAL",
          "NO_PUBLICABLE",
          "SIN_RESULTADOS"
        ],
        "title": "DataStatus",
        "description": "Qué tan sostenible es lo que devuelve la respuesta."
      },
      "Evidencia": {
        "properties": {
          "evidencia_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidencia Id"
          },
          "fragmento": {
            "type": "string",
            "title": "Fragmento"
          },
          "norma": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Norma"
          },
          "unidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidad"
          },
          "url_fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Fuente"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "capturado_en": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capturado En"
          }
        },
        "type": "object",
        "required": [
          "evidencia_id",
          "fragmento"
        ],
        "title": "Evidencia",
        "description": "Localizador verificable de una afirmación."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NormaDetalle": {
        "properties": {
          "norma": {
            "$ref": "#/components/schemas/NormaResumen"
          },
          "versiones": {
            "items": {
              "$ref": "#/components/schemas/VersionNorma"
            },
            "type": "array",
            "title": "Versiones"
          },
          "campos": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Campos"
          },
          "relaciones": {
            "items": {
              "$ref": "#/components/schemas/RelacionResumen"
            },
            "type": "array",
            "title": "Relaciones"
          },
          "referencias_pendientes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Referencias Pendientes"
          }
        },
        "type": "object",
        "required": [
          "norma"
        ],
        "title": "NormaDetalle"
      },
...
      "Pagina_NormaResumen_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NormaResumen"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limite": {
            "type": "integer",
            "title": "Limite"
          },
          "desplazamiento": {
            "type": "integer",
            "title": "Desplazamiento"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limite",
          "desplazamiento"
        ],
        "title": "Pagina[NormaResumen]"
      },
...
      "Respuesta_NormaDetalle_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "known_at": {
            "type": "string",
            "format": "date-time",
            "title": "Known At"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "data": {
            "$ref": "#/components/schemas/NormaDetalle"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidencia"
            },
            "type": "array",
            "title": "Evidence"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Advertencia"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "known_at",
          "data_status",
          "data"
        ],
        "title": "Respuesta[NormaDetalle]"
      },
...
      "Respuesta_Pagina_NormaResumen__": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "known_at": {
            "type": "string",
            "format": "date-time",
            "title": "Known At"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "data": {
            "$ref": "#/components/schemas/Pagina_NormaResumen_"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidencia"
            },
            "type": "array",
            "title": "Evidence"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Advertencia"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "known_at",
          "data_status",
          "data"
        ],
        "title": "Respuesta[Pagina[NormaResumen]]"
      },
...
      "SolicitudRelease": {
        "properties": {
          "motivo": {
            "type": "string",
            "minLength": 1,
            "title": "Motivo"
          },
          "candidatos": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidatos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "SolicitudRelease"
      },
      "SolicitudResolucion": {
        "properties": {
          "decision": {
            "type": "string",
            "minLength": 1,
            "title": "Decision"
          },
          "fundamento_evidencia_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fundamento Evidencia Id"
          },
          "vigencia": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Vigencia"
              },
              {
                "type": "null"
              }
            ]
          },
          "estado_esperado": {
            "type": "string",
            "title": "Estado Esperado",
            "default": "ABIERTA"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "SolicitudResolucion"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}